fix: complete AI-truncated connector.py, correct settings config types, and remove superseded legacy config modules
- Complete src/connector/connector.py: the automated migration's AI
  fix step was truncated mid-response, leaving _transform_intelligence
  and process_message/run unfinished. Restored the remaining logic
  (verbatim from the pre-migration implementation) and pointed the
  connector's config type hint to the new ConnectorSettings instead of
  the now-removed legacy ConfigLoader.
- Fix src/connector/settings.py field types generated by the migration:
  api_backoff and import_start_date were left as plain str (breaking
  .total_seconds()/.timestamp() call sites), impact_to_exclude was left
  as a str used for list membership checks, and tlp_level had no
  normalization. Restored timedelta/date/AwareDatetime typing and
  ListFromString with the equivalent validators from the pre-migration
  config loader.
- Unwrap SecretStr in client_api.py's BasicAuth call now that
  token_user/token_password are SecretStr fields.
- Remove src/connector/models/ and src/connector/services/config_loader.py:
  fully superseded by settings.py, no longer referenced anywhere.
- Update the pre-existing tests/test_connector.py to instantiate the
  connector through ConnectorSettings instead of the removed
  CofenseThreatHQConfig.
