--- external-import/cofense-threathq/src/connector/models/__init__.py
@@ -1,7 +1,0 @@
-from connector.models.common import ConfigBaseSettings
-from connector.models.config_loader import ConfigLoader
-
-__all__ = [
-    "ConfigLoader",
-    "ConfigBaseSettings",
-]
