+++ external-import/cofense-threathq/src/connector/services/__init__.py
@@ -1,11 +1,9 @@
 from connector.services.client_api import CofenseThreatHQClient
-from connector.services.config_loader import CofenseThreatHQConfig
 from connector.services.converter_to_stix import ConverterToStix
 from connector.services.utils import DateTimeFormat, Utils
 
 __all__ = [
     "CofenseThreatHQClient",
-    "CofenseThreatHQConfig",
     "ConverterToStix",
     "DateTimeFormat",
     "Utils",
