+++ external-import/cofense-threathq/src/connector/settings.py
@@ -1,11 +1,13 @@
-from datetime import timedelta
-from pydantic import Field, SecretStr
+from datetime import date, datetime, timedelta, timezone
+from typing import Literal
+
 from connectors_sdk import (
     BaseConfigModel,
     BaseConnectorSettings,
     BaseExternalImportConnectorConfig,
     ListFromString,
 )
+from pydantic import AwareDatetime, Field, SecretStr, field_validator
 
 
 class ExternalImportConnectorConfig(BaseExternalImportConnectorConfig):
@@ -43,8 +45,8 @@ class CofenseThreathqConfig(BaseConfigModel):
     token_password: SecretStr = Field(
         description="Token password for Cofense ThreatHQ API authentication.",
     )
-    import_start_date: str = Field(
-        default="P30D",
+    import_start_date: date | AwareDatetime | timedelta = Field(
+        default=timedelta(days=30),
         description="Date from which data import should start (YYYY-MM-DD, full datetime, or ISO 8601 duration like P30D).",
     )
     api_base_url: str = Field(
@@ -63,19 +65,19 @@ class CofenseThreathqConfig(BaseConfigModel):
         default=5,
         description="Maximum number of retry attempts in case of API failure.",
     )
-    api_backoff: str = Field(
-        default="PT30S",
-        description="Exponential backoff duration between API retries (ISO 8601 format).",
+    api_backoff: timedelta = Field(
+        default=timedelta(seconds=30),
+        description="Exponential backoff duration between API retries (ISO 8601 duration format).",
     )
-    impact_to_exclude: str = Field(
-        default="",
+    impact_to_exclude: ListFromString = Field(
+        default=[],
         description="Comma-separated list of report impacts to exclude (e.g., None,Moderate,Major).",
     )
     import_report_pdf: bool = Field(
         default=True,
         description="Retrieve and attach PDF reports generated by Cofense ThreatHQ.",
     )
-    tlp_level: str = Field(
+    tlp_level: Literal["clear", "green", "amber", "amber+strict", "red"] = Field(
         default="amber+strict",
         description="TLP marking for exported data (clear, green, amber, amber+strict, red).",
     )
@@ -83,6 +85,26 @@ class CofenseThreathqConfig(BaseConfigModel):
         default=True,
         description="Create indicators based on observables from Cofense ThreatHQ.",
     )
+
+    @field_validator("impact_to_exclude", mode="after")
+    @classmethod
+    def _lowercase_impact_to_exclude(cls, value: list[str]) -> list[str]:
+        return [item.strip().lower() for item in value if item.strip()]
+
+    @field_validator("tlp_level", mode="before")
+    @classmethod
+    def _lowercase_tlp_level(cls, value: str) -> str:
+        return value.lower() if isinstance(value, str) else value
+
+    @field_validator("import_start_date", mode="after")
+    @classmethod
+    def _convert_import_start_date_relative_to_utc_datetime(
+        cls, value: date | AwareDatetime | timedelta
+    ) -> date | AwareDatetime | datetime:
+        """Allow relative import_start_date values (timedelta)."""
+        if isinstance(value, timedelta):
+            return datetime.now(tz=timezone.utc) - value
+        return value
 
 
 class ConnectorSettings(BaseConnectorSettings):
@@ -93,4 +115,6 @@ class ConnectorSettings(BaseConnectorSettings):
     connector: ExternalImportConnectorConfig = Field(
         default_factory=ExternalImportConnectorConfig
     )
-    cofense_threathq: CofenseThreathqConfig = Field(default_factory=CofenseThreathqConfig)
+    cofense_threathq: CofenseThreathqConfig = Field(
+        default_factory=CofenseThreathqConfig
+    )
